separating modals
